// src/erp-ui/projects/api/src/lib/model/custom-field-type.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum CustomFieldType {
    Text = 'Text',
    Number = 'Number',
    Date = 'Date',
    Boolean = 'Boolean',
    Select = 'Select',
    MultiSelect = 'MultiSelect'
}